Auth: Add JWT authentication and authorization.
This commits adds necessary logic for authentication and authorization.
Created separate class RequestAccount for handling requests (because
Account class is abstract, therefore it can't have an object instance).
A request to be authorized needs an Authorization header with JWT and
'Bearer ' prefix.
Also implemented a simple endpoint in ExampleController that works only
for users with STUDENT role. This should later be either removed or
moved to a StudentController or AccountController depending whether such
endpoint will be needed or not.